refactor: Add validation to methods and claz
Still working on it.
