+++ src/main/java/io/ib67/astralflow/api/AstralFlowAPI.java
@@ -21,6 +21,7 @@
 
 package io.ib67.astralflow.api;
 
+import com.google.gson.annotations.Since;
 import io.ib67.astralflow.config.AstralFlowConfiguration;
 import io.ib67.astralflow.hook.HookType;
 import io.ib67.astralflow.hook.event.HookEvent;
@@ -31,10 +32,12 @@ import io.ib67.astralflow.manager.ITickManager;
 import io.ib67.astralflow.manager.ItemRegistry;
 import io.ib67.astralflow.texture.ITextureRegistry;
 import org.bukkit.plugin.Plugin;
+import org.jetbrains.annotations.ApiStatus;
 
 import java.util.Collection;
 import java.util.function.Consumer;
 
+@ApiStatus.AvailableSince("0.1.0")
 public interface AstralFlowAPI {
     IMachineManager getMachineManager();
